+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### API Breaking
+
+* [#228](https://github.com/babylonlabs-io/babylon/pull/228) Add inclusion height to early unbonding event
+
 ## v0.14.0
 
 ### State Machine Breaking